+++ NEWS.md
@@ -3,6 +3,6 @@
 
 ## Minor changes
 * Fixed some types in the documentation
-
+* New arguments to simulate.plcp_multi; save_folder, etc